fix: resolve set main and walking value not updating immediately
