+++ src/components/SwitchActive.jsx
@@ -1,12 +1,10 @@
 /* eslint-disable no-unused-vars */
-import React, { useContext } from 'react';
+import React, { useContext, useEffect } from 'react';
 import { styled } from '@mui/material/styles';
 import Paper from '@mui/material/Paper';
 import Grid from '@mui/material/Grid';
 import Switch from '@mui/material/Switch';
 import Typography from '@mui/material/Typography';
-
-import { useHandleProcess, useLogger, usePublisher } from 'kumo-app';
 
 import WalkContext from '../context/WalkContext';
 
@@ -32,42 +30,6 @@ function SwitchState({ name, value, type }) {
   const {
     main, kinematic, walking, setMainValue, setWalkingValue,
   } = useContext(WalkContext);
-  const configPublisher = usePublisher();
-  const walkConfigPublisher = usePublisher();
-  const logger = useLogger();
-
-  const [publishingWalking, handlePublishWalking] = useHandleProcess(() => {
-    const run = main.start;
-    const x_move = main.x;
-    const y_move = main.y;
-    const a_move = main.a;
-    const aim_on = main.aim;
-    return walkConfigPublisher
-      .publish({
-        run, x_move, y_move, a_move, aim_on,
-      })
-      .then(() => {
-        logger.success('Successfully publish main config.');
-      })
-      .catch((err) => {
-        logger.error(`Failed to publish main config! ${err.message}.`);
-      });
-  }, 500);
-
-  const [publishingConfig, handlePublishConfig] = useHandleProcess(() => {
-    const json_kinematic = JSON.stringify(kinematic);
-    const json_walking = JSON.stringify(walking);
-    return configPublisher
-      .publish({
-        json_kinematic, json_walking,
-      })
-      .then(() => {
-        logger.success('Successfully publish kinematic and walking config.');
-      })
-      .catch((err) => {
-        logger.error(`Failed to publish kinematic and walking config! ${err.message}.`);
-      });
-  }, 500);
 
   return (
     <Item>
@@ -84,11 +46,9 @@ function SwitchState({ name, value, type }) {
           <Switch
             onChange={type === 'main' ? (event) => {
               setMainValue(name, event.target.checked);
-              handlePublishWalking();
             }
               : (event) => {
                 setWalkingValue('balance', name, event.target.checked);
-                handlePublishConfig();
               }}
             checked={value}
           />
